requirements.txt
streamlit
pandas
plotly
groq
openpyxl
Pillow
